core / refactor / collect and export all interfaces
