+++ goodmap-core/src/index.ts
@@ -1,5 +1,7 @@
 import GoodmapCore from './GoodmapCore';
 
+export * from './interfaces';
+
 export const {
   createEntry,
   getEntriesByIds,
