+++ goodmap-core/src/interfaces/index.ts
@@ -1,0 +1,8 @@
+import IEntry from './IEntry';
+import ISearchResult from './ISearchResult';
+import IUserCredentials from './IUserCredentials';
+import IUserProfile from './IUserProfile';
+
+export {
+  IEntry, ISearchResult, IUserCredentials, IUserProfile,
+};
